requirements.txt
# --- NLP core ---
spacy==3.7.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1.tar.gz

# --- LangChain ecosystem ---
langchain==0.2.16
langchain-google-vertexai==1.0.6
langchain-community==0.2.16

# --- CrewAI market analyst agent ---
crewai>=0.80.0
litellm>=1.0.0
ddgs>=1.0.0

# --- LangGraph agent ---
langgraph>=1.0.0

# --- Gemini AI Studio (generative narrative) ---
google-genai>=1.0.0

# --- GCP / Vertex AI ---
google-cloud-aiplatform==1.62.0
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
google-cloud-monitoring>=2.0.0
google-auth==2.32.0

# --- API layer ---
fastapi==0.111.1
uvicorn[standard]==0.30.3
python-multipart==0.0.9

# --- PDF ingestion ---
pdfplumber==0.11.4

# --- Data & validation ---
pydantic==2.8.2
pydantic-settings==2.4.0

# --- Utilities ---
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.5.0

# --- Vertex AI Pipelines (KFP v2) ---
kfp==2.8.0

# --- Dev / test ---
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
